+++ http/package.json
@@ -36,7 +36,9 @@
   },
   "scripts": {
     "build": "tsc --build tsconfig.build.json && pnpm run build:cjs && pnpm run build:esm",
-    "build:cjs": "BUNDLE_FORMAT=cjs rollup -c ../rollup.config.ts",
-    "build:esm": "BUNDLE_FORMAT=esm rollup -c ../rollup.config.ts"
-  }
+    "build:cjs": "BUNDLE_FORMAT=cjs rollup -c ../configs/rollup.config.ts",
+    "build:esm": "BUNDLE_FORMAT=esm rollup -c ../configs/rollup.config.ts",
+    "format": "prettier --write './{src,test}/**/*'"
+  },
+  "prettier": "../configs/.prettierrc.json"
 }
