+++ http/src/run.ts
@@ -1,6 +1,6 @@
 // This is all the stuff that should probably be part of @cycle/run
 
-import { Source } from "@cycle/callbags";
+import { Source } from '@cycle/callbags';
 
 export type Subscription = void;
