+++ http/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../tsconfig.json",
+  "extends": "../configs/tsconfig.json",
   "include": [
     "src/**/*"
   ]
